backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic[email]==2.12.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
bcrypt==4.1.3
starlette==0.37.2
email-validator==2.1.0
